requirements.txt
cassandra-driver
Flask==1.1.2
Jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.2
